+++ R/assertthat.R
@@ -349,3 +349,28 @@ on_failure(has_tabletree_colnames) <- function(call, env) {
 
   paste0("required column names ", deparse(missing_col_names), " are not found in table")
 }
+
+#' @describeIn assertions check if the input `df` has `na_level` in its `variables`.
+#' @param df (`data frame`)\cr input dataset.
+#' @param variables (`named list`)\cr columns from dataset where you want to check if `na_level` exists.
+#' @param na_level (`string`)\cr the string user has been using to represent NA or missing data.
+#' @export
+#'
+#' @examples
+#'
+#' df <- data.frame(a = 1:3, b = 2:4)
+#' df$a <- ifelse(df$a == 1, "<Missing>", df$a)
+#' is_df_with_no_na_level(df, variables = list(a = "a"), na_level = "<Missing>")
+#' is_df_with_no_na_level(df, variables = list(b = "b"), na_level = "<Missing>")
+#' is_df_with_no_na_level(df, variables = list(a = "a", b = "b"), na_level = "<Missing>")
+#'
+is_df_with_no_na_level <- function(df, variables, na_level) {
+  assert_that(
+    is_df_with_variables(df, variables),
+    is.string(na_level)
+  )
+  !any(df[, unlist(variables)] == na_level)
+}
+on_failure(is_df_with_no_na_level) <- function(call, env) {
+  paste(deparse(call$df), "contains missing data as defined by the argument na_level")
+}